Refactor the Julia package to use a common package for the primal codes
