+++ julia/Gradbench/src/benchmarks/hello.jl
@@ -1,0 +1,7 @@
+module hello
+
+function square(x)
+    return x * x
+end
+
+end
