+++ tools/zygote/Project.toml
@@ -1,3 +1,3 @@
 [deps]
-JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+Gradbench = "b8d9ff47-6a34-4c60-b82b-acd435921e41"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
